server/requirements.txt
# screenpoint
photoshop-connection
Flask
flask-cors
pyscreenshot
pyautogui
pygetwindow
Pillow
requests
itsdangerous
Werkzeug
pywin32
